RMX1921: RealmeParts: Update preferences and Cleanup!
* Update CustomSeekBarPreference, VibrationSeekBarPreference & ProperSeekBarPreference from derp and adapt it for our RealmeParts.
* Commit History: https://github.com/DerpFest-AOSP/vendor_support/commits/13/src/org/derpfest/support/preferences/ProperSeekBarPreference.java
* Cleanup all unused files while at it.
