--- RealmeParts/res/drawable/ic_joystick.xml
@@ -1,23 +1,0 @@
-<vector android:height="24dp" android:viewportHeight="511.997"
-    android:viewportWidth="511.997" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FFD33F" android:pathData="M136.549,247.463c-4.727,0 -9.454,0.256 -14.155,0.759C67.687,254.101 15.454,306.121 9.379,360.845C1.299,431.064 51.681,494.534 121.9,502.614c28.625,3.293 57.515,-3.174 82.01,-18.344c12.406,-8.071 26.56,-13.046 41.287,-14.505c3.549,-0.316 7.15,-0.469 10.802,-0.469c3.464,0 6.902,0.119 10.307,0.367c14.948,1.408 29.333,6.433 41.909,14.641c60.151,37.123 139.004,18.455 176.127,-41.696c15.093,-24.453 21.518,-53.274 18.233,-81.814c-6.066,-54.682 -58.308,-106.693 -113.016,-112.572c-4.684,-0.503 -9.394,-0.759 -14.112,-0.759L136.549,247.463L136.549,247.463z"/>
-    <path android:fillColor="#EA3836" android:pathData="M315.688,375.409m-25.596,0a25.596,25.596 0,1 1,51.192 0a25.596,25.596 0,1 1,-51.192 0"/>
-    <path android:fillColor="#FFAC3E" android:pathData="M478.574,302.656c-102.692,115.089 -333.97,157.92 -424.299,170.641c42.225,35.698 102.649,40.109 149.61,10.93c12.406,-8.071 26.56,-13.046 41.287,-14.504c3.566,-0.282 7.184,-0.427 10.827,-0.427c3.464,0 6.902,0.119 10.307,0.367c14.948,1.408 29.333,6.433 41.909,14.641c60.151,37.123 139.004,18.455 176.127,-41.696c15.093,-24.453 21.518,-53.274 18.233,-81.814C499.827,339.643 491.543,319.593 478.574,302.656z"/>
-    <path android:fillColor="#383856" android:pathData="M119.485,315.719l34.128,0l0,42.66l42.661,0l0,34.129l-42.661,0l0,42.66l-34.128,0l0,-42.66l-42.66,0l0,-34.129l42.66,0z"/>
-    <path android:fillColor="#91CECE" android:pathData="M375.413,324.217m-25.596,0a25.596,25.596 0,1 1,51.192 0a25.596,25.596 0,1 1,-51.192 0"/>
-    <path android:fillColor="#91CECE" android:pathData="M375.413,426.602m-25.596,0a25.596,25.596 0,1 1,51.192 0a25.596,25.596 0,1 1,-51.192 0"/>
-    <path android:fillColor="#EA3836" android:pathData="M435.137,375.409m-25.596,0a25.596,25.596 0,1 1,51.192 0a25.596,25.596 0,1 1,-51.192 0"/>
-    <path android:fillColor="#874D1B" android:pathData="M238.934,247.463l0,-34.129l34.128,0l0,34.129"/>
-    <path android:fillColor="#D62121" android:pathData="M328.981,353.653c2.474,3.984 3.788,8.575 3.805,13.259c0,14.138 -11.459,25.596 -25.596,25.596c-4.684,-0.017 -9.274,-1.331 -13.259,-3.805c7.286,12.064 22.985,15.938 35.05,8.651s15.938,-22.985 8.652,-35.05C335.5,358.763 332.522,355.794 328.981,353.653z"/>
-    <path android:fillColor="#83B2B7" android:pathData="M388.706,302.46c2.474,3.984 3.788,8.575 3.805,13.259c0,14.138 -11.459,25.596 -25.596,25.596c-4.684,-0.017 -9.274,-1.331 -13.259,-3.805c7.286,12.064 22.985,15.938 35.05,8.651c12.064,-7.286 15.938,-22.985 8.651,-35.05C395.224,307.571 392.247,304.602 388.706,302.46z"/>
-    <path android:fillColor="#D62121" android:pathData="M448.43,353.653c2.474,3.984 3.788,8.575 3.805,13.259c0,14.138 -11.459,25.596 -25.596,25.596c-4.684,-0.017 -9.274,-1.331 -13.259,-3.805c7.286,12.064 22.985,15.938 35.05,8.651s15.938,-22.985 8.651,-35.05C454.949,358.763 451.971,355.794 448.43,353.653z"/>
-    <path android:fillColor="#83B2B7" android:pathData="M388.706,404.845c2.474,3.984 3.788,8.575 3.805,13.259c0,14.138 -11.459,25.596 -25.596,25.596c-4.684,-0.017 -9.274,-1.331 -13.259,-3.805c7.286,12.064 22.985,15.938 35.05,8.651c12.064,-7.286 15.938,-22.985 8.651,-35.05C395.224,409.956 392.247,406.986 388.706,404.845z"/>
-    <path android:fillColor="#202033" android:pathData="M76.825,384.172v8.336h42.66v42.66h34.128v-42.66h42.66v-34.128h-14.24C148.332,381.877 96.944,384.231 76.825,384.172z"/>
-    <path android:fillColor="#6B3713" android:pathData="M238.934,245.133v2.329h34.128v-28.028C264.931,231.627 252.901,240.688 238.934,245.133z"/>
-    <path android:fillColor="#FF000000" android:pathData="M196.274,349.847h-34.128v-34.128c0,-4.71 -3.822,-8.532 -8.532,-8.532h-34.128c-4.71,0 -8.532,3.822 -8.532,8.532v34.128H76.825c-4.71,0 -8.532,3.822 -8.532,8.532v34.128c0,4.71 3.822,8.532 8.532,8.532h34.128v34.128c0,4.71 3.822,8.532 8.532,8.532h34.128c4.71,0 8.532,-3.822 8.532,-8.532V401.04h34.128c4.71,0 8.532,-3.822 8.532,-8.532v-34.128C204.806,353.67 200.983,349.847 196.274,349.847zM187.742,383.976h-34.128c-4.71,0 -8.532,3.822 -8.532,8.532v34.128h-17.064v-34.128c0,-4.71 -3.822,-8.532 -8.532,-8.532H85.357v-17.064h34.128c4.71,0 8.532,-3.822 8.532,-8.532v-34.128h17.064v34.128c0,4.71 3.822,8.532 8.532,8.532h34.128V383.976z"/>
-    <path android:fillColor="#FF000000" android:pathData="M375.447,358.379c18.847,0 34.128,-15.281 34.128,-34.128c0,-18.847 -15.281,-34.128 -34.128,-34.128c-18.847,0 -34.128,15.281 -34.128,34.128C341.319,343.098 356.6,358.379 375.447,358.379zM375.447,307.187c9.428,0 17.064,7.636 17.064,17.064s-7.636,17.064 -17.064,17.064s-17.064,-7.636 -17.064,-17.064S366.028,307.187 375.447,307.187z"/>
-    <path android:fillColor="#FF000000" android:pathData="M375.447,392.508c-18.847,0 -34.128,15.281 -34.128,34.128c0,18.847 15.281,34.128 34.128,34.128c18.847,0 34.128,-15.281 34.128,-34.128C409.575,407.788 394.294,392.508 375.447,392.508zM375.447,443.7c-9.428,0 -17.064,-7.636 -17.064,-17.064s7.636,-17.064 17.064,-17.064s17.064,7.636 17.064,17.064S384.875,443.7 375.447,443.7z"/>
-    <path android:fillColor="#FF000000" android:pathData="M315.723,341.315c-18.847,0 -34.128,15.281 -34.128,34.128c0,18.847 15.281,34.128 34.128,34.128c18.847,0 34.128,-15.281 34.128,-34.128C349.851,356.596 334.57,341.315 315.723,341.315zM315.723,392.508c-9.428,0 -17.064,-7.636 -17.064,-17.064s7.636,-17.064 17.064,-17.064s17.064,7.636 17.064,17.064S325.15,392.508 315.723,392.508z"/>
-    <path android:fillColor="#FF000000" android:pathData="M435.171,341.315c-18.847,0 -34.128,15.281 -34.128,34.128c0,18.847 15.281,34.128 34.128,34.128s34.128,-15.281 34.128,-34.128C469.299,356.596 454.019,341.315 435.171,341.315zM435.171,392.508c-9.428,0 -17.064,-7.636 -17.064,-17.064s7.636,-17.064 17.064,-17.064s17.064,7.636 17.064,17.064S444.599,392.508 435.171,392.508z"/>
-    <path android:fillColor="#FF000000" android:pathData="M390.497,239.784c-5,-0.555 -10.025,-0.836 -15.051,-0.853h-93.853v-34.128H264.53v-17.064c0,-18.847 -15.281,-34.128 -34.128,-34.128h-42.66c-9.428,0 -17.064,-7.636 -17.064,-17.064s7.636,-17.064 17.064,-17.064h136.513c18.847,0 34.128,-15.281 34.128,-34.128s-15.281,-34.128 -34.128,-34.128h-42.66c-9.428,0 -17.064,-7.636 -17.064,-17.064V0.033h-17.064v34.128c0,18.847 15.281,34.128 34.128,34.128h42.66c9.428,0 17.064,7.636 17.064,17.064s-7.636,17.064 -17.064,17.064H187.742c-18.847,0 -34.128,15.281 -34.128,34.128s15.281,34.128 34.128,34.128h42.66c9.428,0 17.064,7.636 17.064,17.064v17.064h-17.064v34.128h-93.853c-5.034,0.009 -10.068,0.299 -15.068,0.853C62.611,246.098 7.425,301.104 0.898,359.898c-8.575,74.903 45.203,142.579 120.106,151.145c30.502,3.49 61.294,-3.404 87.394,-19.564c11.237,-7.355 24.086,-11.894 37.447,-13.233c6.604,-0.52 13.233,-0.546 19.837,-0.068c13.575,1.314 26.629,5.904 38.036,13.378c64.135,39.64 148.253,19.777 187.893,-44.358c16.126,-26.1 22.994,-56.866 19.496,-87.343C504.563,301.061 449.343,246.055 390.497,239.784zM247.466,221.866h17.064v17.064h-17.064V221.866zM464.479,455.082c-38.573,43.147 -102.538,52.404 -151.768,21.97c-13.745,-8.95 -29.47,-14.411 -45.809,-15.895c-3.601,-0.264 -7.235,-0.392 -10.904,-0.392c-3.882,0 -7.679,0.162 -11.501,0.461c-16.091,1.544 -31.569,6.945 -45.118,15.75c-56.098,34.708 -129.713,17.371 -164.43,-38.727c-14.129,-22.84 -20.153,-49.767 -17.09,-76.456c5.614,-50.527 54.878,-99.603 105.456,-105.038c4.394,-0.495 8.814,-0.742 13.233,-0.759h238.898c4.42,0 8.839,0.239 13.233,0.708c50.57,5.435 99.825,54.511 105.456,105.038C498.163,395.673 487.353,429.69 464.479,455.082z"/>
-</vector>
