--- RealmeParts/res/drawable/ic_preset.xml
@@ -1,15 +1,0 @@
-<vector android:height="24dp" android:viewportHeight="464"
-    android:viewportWidth="464.01605" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#4db7e5" android:pathData="m367.703,43.813c-92.32,-66.543 -219.277,-56.32 -299.762,24.137 -0.961,0.961 -1.84,1.961 -2.781,2.93l61.848,3.633 3.27,55.582c46.133,-45.992 117.398,-55.27 173.77,-22.625l-26.105,26.145 56,-3.285 55.586,-3.273 3.633,-61.84 2.926,-49.719zM367.703,43.813"/>
-    <path android:fillColor="#4db7e5" android:pathData="m337.008,389.52 l-3.273,-55.578c-46.125,45.973 -117.371,55.246 -173.727,22.617l26.102,-26.105 -56,3.289 -55.582,3.273 -3.633,61.84 -2.93,49.727 28.371,-28.367c92.32,66.543 219.273,56.32 299.758,-24.137 0.953,-0.961 1.84,-1.953 2.777,-2.93zM337.008,389.52"/>
-    <path android:fillColor="#348fd9" android:pathData="m107.461,160.016 l26.145,26.063 -3.289,-56 -3.27,-55.582 -61.848,-3.633 -49.754,-2.914 28.359,28.371c-66.543,92.32 -56.32,219.273 24.137,299.758 0.961,0.961 1.961,1.84 2.93,2.777l3.633,-61.84 55.582,-3.273c-45.977,-46.121 -55.254,-117.367 -22.625,-173.727zM107.461,160.016"/>
-    <path android:fillColor="#348fd9" android:pathData="m396.07,67.949c-0.961,-0.949 -1.961,-1.84 -2.93,-2.773l-3.629,61.84 -55.586,3.27c45.996,46.129 55.277,117.402 22.625,173.77l-26.105,-26.105 3.289,56 3.273,55.578 61.84,3.633 49.719,2.926 -28.359,-28.367c66.547,-92.324 56.324,-219.281 -24.137,-299.77zM396.07,67.949"/>
-    <path android:fillColor="#ffb531" android:pathData="m320.008,248.016v-32h-17.848c-1.801,-7.898 -4.93,-15.43 -9.25,-22.281l12.633,-12.633 -22.625,-22.625 -12.633,12.633c-6.848,-4.32 -14.383,-7.449 -22.277,-9.246v-17.848h-32v17.848c-7.898,1.797 -15.434,4.926 -22.281,9.246l-12.633,-12.633 -22.625,22.625 12.633,12.633c-4.32,6.852 -7.449,14.383 -9.246,22.281h-17.848v32h17.848c1.797,7.895 4.926,15.43 9.246,22.277l-12.633,12.633 22.625,22.625 12.633,-12.633c6.848,4.32 14.383,7.449 22.281,9.25v17.848h32v-17.848c7.895,-1.801 15.43,-4.93 22.277,-9.25l12.633,12.633 22.625,-22.625 -12.633,-12.633c4.32,-6.848 7.449,-14.383 9.25,-22.277zM232.008,264.016c-17.676,0 -32,-14.328 -32,-32 0,-17.676 14.324,-32 32,-32 17.672,0 32,14.324 32,32 0,17.672 -14.328,32 -32,32zM232.008,264.016"/>
-    <path android:fillColor="#62a9e6" android:pathData="m54.293,123.84c-3.313,5.406 -6.398,11.008 -9.184,16.629l14.344,7.09c2.563,-5.199 5.418,-10.398 8.48,-15.367zM54.293,123.84"/>
-    <path android:fillColor="#62a9e6" android:pathData="m37.773,157.109c-14.469,37.066 -17.902,77.531 -9.879,116.504l15.672,-3.199c-7.402,-35.961 -4.234,-73.293 9.113,-107.496zM37.773,157.109"/>
-    <path android:fillColor="#ffa912" android:pathData="m292.91,193.734 l12.633,-12.633 -11.313,-11.313 -39.602,39.602c8.32,8.035 11.656,19.934 8.727,31.121 -2.926,11.188 -11.664,19.926 -22.852,22.852 -11.191,2.93 -23.086,-0.406 -31.121,-8.727l-39.602,39.602 11.313,11.313 12.633,-12.633c6.848,4.32 14.383,7.449 22.281,9.25v17.848h32v-17.848c7.895,-1.801 15.43,-4.93 22.277,-9.25l12.633,12.633 22.625,-22.625 -12.633,-12.633c4.32,-6.848 7.449,-14.383 9.25,-22.277h17.848v-32h-17.848c-1.801,-7.898 -4.93,-15.43 -9.25,-22.281zM292.91,193.734"/>
-    <path android:fillColor="#2d7dbf" android:pathData="m401.445,137.453 l3.512,-59.766c-2.918,-3.266 -5.75,-6.594 -8.887,-9.738 -0.961,-0.949 -1.961,-1.84 -2.93,-2.773l-3.629,61.84 -55.586,3.27c3.391,3.395 6.512,6.945 9.496,10.578l2.449,-0.145zM401.445,137.453"/>
-    <path android:fillColor="#45a7d1" android:pathData="m398.848,393.148 l-61.84,-3.629 -3.273,-55.586c-3.391,3.395 -6.945,6.52 -10.574,9.496l0.141,2.449 3.266,55.574 59.77,3.512c3.27,-2.918 6.59,-5.75 9.734,-8.887 0.961,-0.961 1.84,-1.953 2.777,-2.93zM398.848,393.148"/>
-    <path android:fillColor="#45a7d1" android:pathData="m81.16,58.941 l-3.137,-0.191c-3.395,3.016 -6.832,5.953 -10.082,9.199 -0.961,0.961 -1.84,1.961 -2.781,2.93l61.848,3.633 3.27,55.582c4.965,-4.914 10.277,-9.457 15.898,-13.598l-3.203,-53.922zM81.16,58.941"/>
-    <path android:fillColor="#2d7dbf" android:pathData="m70.871,398.855 l3.633,-61.84 55.582,-3.273c-3.391,-3.391 -6.52,-6.945 -9.496,-10.574l-2.449,0.141 -55.574,3.266 -3.512,59.77c2.918,3.27 5.75,6.59 8.887,9.734 0.961,0.961 1.961,1.84 2.93,2.777zM70.871,398.855"/>
-</vector>
