--- RealmeParts/res/drawable/ic_swipe.xml
@@ -1,15 +1,0 @@
-<vector android:height="24dp" android:viewportHeight="512"
-    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#583377" android:pathData="m511.219,255.609c0,141.168 -114.441,255.609 -255.609,255.609 -2.965,0 -5.906,-0.055 -8.836,-0.156 -63.141,-2.141 -120.461,-27.18 -163.934,-67.086 -50.926,-46.719 -82.84,-113.813 -82.84,-188.367 0,-141.168 114.441,-255.609 255.609,-255.609s255.609,114.441 255.609,255.609zM511.219,255.609"/>
-    <path android:fillColor="#5a9ed1" android:pathData="m466.422,84.777c-86.168,-86.168 -226.371,-86.168 -312.539,0 -41.738,41.742 -64.723,97.223 -64.723,156.238v0.035c0,31.75 16.258,160.07 130.496,181.98 114.234,21.906 -63.883,-181.988 -63.883,-181.988 0,-0.012 0,-0.016 0,-0.023 0,-41.223 16.055,-79.98 45.211,-109.137 60.195,-60.195 158.141,-60.195 218.336,0l-45.582,45.578h138.262v-138.262zM466.422,84.777"/>
-    <path android:fillColor="#ff4f9b" android:pathData="m272.668,440.023 l-25.895,71.039c-63.141,-2.141 -120.461,-27.18 -163.934,-67.086l24.922,-68.355 7.773,3.035zM272.668,440.023"/>
-    <path android:fillColor="#f9d8c7" android:pathData="m272.668,440.02 l-12.68,34.785s-68.461,19.34 -132.176,-21.547c-12.605,-8.094 -25.023,-18.539 -36.695,-31.969l16.648,-45.668 7.77,3.035zM272.668,440.02"/>
-    <path android:fillColor="#f9d8c7" android:pathData="m272.668,440.02 l-12.68,34.785s-68.461,19.34 -132.176,-21.547c-12.605,-8.094 -25.023,-18.539 -36.695,-31.969l16.648,-45.668 7.77,3.035zM272.668,440.02"/>
-    <path android:fillColor="#e0a792" android:pathData="m142.441,413.129 l-14.629,40.129c-12.605,-8.094 -25.023,-18.539 -36.695,-31.965l16.648,-45.668 7.77,3.031zM142.441,413.129"/>
-    <path android:fillColor="#f9d8c7" android:pathData="m371.07,326.523c-12.234,10.629 -23.133,40.488 -23.195,40.66 0,0 -42.527,114.953 -125.691,84.633 -31.082,-11.332 -74.863,-27.293 -74.863,-27.293 -61.375,-22.375 -73.145,-97.832 -53.438,-151.887l21.137,-57.969c2.523,-6.93 7.629,-12.199 13.836,-15.09 6.207,-2.879 13.516,-3.41 20.457,-0.879 13.887,5.063 21.039,20.426 15.98,34.301l9.164,-25.137c2.531,-6.941 7.629,-12.199 13.84,-15.09 6.207,-2.891 13.512,-3.41 20.457,-0.879 6.93,2.527 12.199,7.629 15.09,13.84 2.879,6.207 3.41,13.516 0.879,20.457l2.336,-6.41c2.531,-6.941 7.629,-12.199 13.84,-15.09 6.207,-2.891 13.512,-3.41 20.457,-0.879 13.875,5.059 21.027,20.422 15.969,34.293l32.441,-88.98c5.066,-13.863 20.414,-21.008 34.289,-15.949 6.93,2.527 12.188,7.625 15.074,13.844 2.891,6.211 3.422,13.52 0.895,20.449l-49.156,134.828s4.297,-8.125 20.172,-21.902c12.734,-11.066 32.02,-9.699 43.082,3.047 11.059,12.746 9.691,32.031 -3.051,43.082zM371.07,326.523"/>
-    <path android:fillColor="#e0a792" android:pathData="m140.566,418.273 l-1.086,2.98c-54.508,-25.723 -64.426,-96.957 -45.59,-148.621l21.129,-57.961c2.531,-6.938 7.625,-12.199 13.836,-15.102 6.211,-2.875 13.508,-3.402 20.461,-0.867 2.734,0.996 5.215,2.391 7.379,4.102 -4.234,-0.145 -8.41,0.73 -12.168,2.48 -6.207,2.898 -11.313,8.156 -13.844,15.098l-21.129,57.957c-17.016,46.668 -10.566,109.297 31.012,139.934zM140.566,418.273"/>
-    <path android:fillColor="#e0a792" android:pathData="m168.34,248.98c-4.324,-1.578 -6.555,-6.363 -4.977,-10.688l11.09,-30.426c2.527,-6.934 7.633,-12.199 13.84,-15.09 6.207,-2.883 13.516,-3.41 20.457,-0.883 2.734,1 5.215,2.398 7.387,4.109 -4.246,-0.145 -8.414,0.734 -12.172,2.484 -6.211,2.891 -11.313,8.16 -13.84,15.09l-11.094,30.426c-1.578,4.328 -6.363,6.555 -10.691,4.977zM168.34,248.98"/>
-    <path android:fillColor="#e0a792" android:pathData="m218.996,266.238c-4.328,-1.574 -6.555,-6.363 -4.98,-10.688l13.043,-35.766c2.523,-6.934 7.629,-12.199 13.836,-15.09 6.207,-2.883 13.516,-3.414 20.457,-0.883 2.734,0.996 5.215,2.398 7.387,4.109 -4.242,-0.145 -8.41,0.734 -12.172,2.484 -6.207,2.891 -11.313,8.156 -13.84,15.09l-13.039,35.766c-1.578,4.328 -6.363,6.555 -10.691,4.977zM218.996,266.238"/>
-    <path android:fillColor="#e0a792" android:pathData="m269.727,283.297c-4.328,-1.578 -6.559,-6.363 -4.98,-10.691l45.02,-123.48c2.527,-6.93 7.629,-12.199 13.84,-15.09 6.203,-2.879 13.512,-3.41 20.457,-0.879 2.734,0.996 5.211,2.395 7.383,4.105 -4.242,-0.141 -8.41,0.734 -12.172,2.484 -6.207,2.891 -11.313,8.16 -13.84,15.09l-45.016,123.48c-1.582,4.328 -6.367,6.559 -10.691,4.98zM269.727,283.297"/>
-    <path android:fillColor="#e0a792" android:pathData="m374.121,283.441c-19.402,-13.797 -45.883,21.559 -62.102,62.148 -1.586,3.965 -5.973,6.016 -10.012,4.621 -4.199,-1.453 -6.395,-6.066 -4.871,-10.242l13.734,-37.668s4.293,-8.129 20.172,-21.906c12.73,-11.066 32.016,-9.699 43.078,3.047zM374.121,283.441"/>
-</vector>
