+++ RealmeParts/res/values/proper_seekbar_strings.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
   <string name="proper_seekbar_value">Value: <xliff:g id="v">%s</xliff:g></string>
-  <string name="proper_seekbar_default_value">by default</string>
-  <string name="proper_seekbar_default_value_to_set">Default value: <xliff:g id="v">%s</xliff:g>\nLong tap to set</string>
+  <string name="proper_seekbar_default_value">Default</string>
+  <string name="proper_seekbar_default_value_to_set">Default value: <xliff:g id="v">%s</xliff:g>\nLong press to set</string>
   <string name="proper_seekbar_default_value_is_set">Default value is set</string>
 </resources>
