+++ RealmeParts/res/values/strings.xml
@@ -25,22 +25,21 @@
     <string name="dc_mode_summary">Reduces screen flicker while brightness is low</string>
     <string name="dc_mode_title">DC-Dimming</string>
     <string name="tile_dc_mode">DC-Dimming</string>
+
     <!-- HBM mode -->
     <string name="hbm_mode_summary">Peak brightness for better visibility in outdoors</string>
     <string name="hbm_mode_title">High brightness mode</string>
-    <string name="tile_hbm_mode">High brightness modes</string>
+    <string name="tile_hbm_mode">High brightness mode</string>
     <string name="hbm_autobrightness_summary"></string>
     <string name="hbm_autobrightness_title">Automatic High brightness mode</string>
     <string name="hbm_autobrightness_threshold_summary"></string>
     <string name="hbm_autobrightness_threshold_title">Automatic High brightness mode threshold (lux)</string>
-    <!-- srgb mode -->
-    <string name="srgb_mode_summary">sRGB mode for simpler colour display</string>
-    <string name="srgb_mode_title">sRGB mode</string>
-    <string name="tile_srgb_mode">sRGB</string>
+
     <!-- otg mode -->
     <string name="otg_mode_summary">Enable USB-OTG power supply for external headphones and flash drives</string>
     <string name="otg_mode_title">Enable USB-OTG</string>
     <string name="tile_otg_mode">Enable USB-OTG</string>
+
     <!-- game mode -->
     <string name="game_mode_summary">Boost Touch Sampling Rate</string>
     <string name="game_mode_title">Game Mode</string>
@@ -162,15 +161,8 @@
 
     <!-- Audio -->
     <string name="audio_title">Audio</string>
-
-    <!-- Custom seekbar -->
-    <string name="custom_seekbar_value">Value: <xliff:g id="v">%s</xliff:g></string>
-    <string name="custom_seekbar_default_value">Default</string>
-    <string name="custom_seekbar_default_value_to_set">Default value: <xliff:g id="v">%s</xliff:g>\nLong press to set</string>
-    <string name="custom_seekbar_default_value_is_set">Default value is set</string>
     <string name="switch_bar_on">On</string>
     <string name="switch_bar_off">Off</string>
-
     <string name="audio_gain_settings">Audio Amplification</string>
     <string name="gain_summary">Improve audio devices gain</string>
